Some small fixes
